Fix header width problem and design buttons
The problem with the header width was fixed such that the header element
would occupy the entire width of its parent container.

The buttons were designed for aesthetic purposes with padding, margin,
and border.

At this point, the project is basically complete, but improvements can
be made to the code to make the webpage responsive. Currently, the page
is not responsive at all.